external/neuronpedia/apps/sparsity/pyproject.toml
[project]
name = "sparsity-server"
version = "0.1.0"
description = "FastAPI server for analyzing MLP neuron connections in sparse circuit models"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "torch",
    "transformers",
    "python-dotenv",
    "circuit-sparsity @ git+https://github.com/openai/circuit_sparsity.git",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "httpx",
]
